add RawLinks Option
This commit introduces type Option; Option values can be invoked to modify parsing. RawLinks is an Option that instructs the parser to output the text of wiki internal links without parsing them.

The default parsing behavior is now to output internal links as the link text displayed on the wiki, with two exceptions:

1. In wiki user signatures, user talk page links and timestamps are skipped.
2. Links to image files within template Card are parsed into the Image field of type Card.